fixed a slight problem with the responses service
will leave the old service there for now until I change the tests, as this will break the tests

created a controller helper folder for controller helpers, moved my new helpers as well as my helpers for campaigns into it
